Refactorizar CreateProduct y Crear hook useCreateProduct
